Crate hpke_rs_crypto

HPKE Crypto Trait

The hpke-rs crate can be used with different cryptography backends to perform the cryptographic operations. This crate defines the HpkeCrypto trait that is required to instantiate hpke-rs. It further defines necessary types and errors.

Please see hpke-rs for more details.

Modules

error

HPKE Crypto Trait Errors

types

HPKE Algorithm Identifiers

Traits

CryptoRng

A marker trait over RngCore for securely unpredictable RNGs

HpkeCrypto

The HpkeCrypto trait defines the necessary cryptographic functions used in the HPKE implementation.

HpkeTestRng

PRNG extension for testing that is supposed to return pre-configured bytes.

RngCore

Implementation-level interface for RNGs