hpke_rs_crypto

Trait HpkeCrypto

pub trait HpkeCrypto:
    Debug
    + Send
    + Sync {
    type HpkePrng: RngCore + CryptoRng + HpkeTestRng;

    // Required methods
    fn name() -> String;
    fn supports_kdf(alg: KdfAlgorithm) -> Result<(), Error>;
    fn supports_kem(alg: KemAlgorithm) -> Result<(), Error>;
    fn supports_aead(alg: AeadAlgorithm) -> Result<(), Error>;
    fn prng() -> Self::HpkePrng;
    fn kdf_extract(
        alg: KdfAlgorithm,
        salt: &[u8],
        ikm: &[u8],
    ) -> Result<Vec<u8>, Error>;
    fn kdf_expand(
        alg: KdfAlgorithm,
        prk: &[u8],
        info: &[u8],
        output_size: usize,
    ) -> Result<Vec<u8>, Error>;
    fn dh(alg: KemAlgorithm, pk: &[u8], sk: &[u8]) -> Result<Vec<u8>, Error>;
    fn secret_to_public(alg: KemAlgorithm, sk: &[u8]) -> Result<Vec<u8>, Error>;
    fn kem_key_gen(
        alg: KemAlgorithm,
        prng: &mut Self::HpkePrng,
    ) -> Result<(Vec<u8>, Vec<u8>), Error>;
    fn kem_key_gen_derand(
        alg: KemAlgorithm,
        seed: &[u8],
    ) -> Result<(Vec<u8>, Vec<u8>), Error>;
    fn kem_encaps(
        alg: KemAlgorithm,
        pk_r: &[u8],
        prng: &mut Self::HpkePrng,
    ) -> Result<(Vec<u8>, Vec<u8>), Error>;
    fn kem_decaps(
        alg: KemAlgorithm,
        ct: &[u8],
        sk_r: &[u8],
    ) -> Result<Vec<u8>, Error>;
    fn dh_validate_sk(alg: KemAlgorithm, sk: &[u8]) -> Result<Vec<u8>, Error>;
    fn aead_seal(
        alg: AeadAlgorithm,
        key: &[u8],
        nonce: &[u8],
        aad: &[u8],
        msg: &[u8],
    ) -> Result<Vec<u8>, Error>;
    fn aead_open(
        alg: AeadAlgorithm,
        key: &[u8],
        nonce: &[u8],
        aad: &[u8],
        msg: &[u8],
    ) -> Result<Vec<u8>, Error>;

    // Provided methods
    fn kdf_digest_length(alg: KdfAlgorithm) -> usize { ... }
    fn aead_key_length(alg: AeadAlgorithm) -> usize { ... }
    fn aead_nonce_length(alg: AeadAlgorithm) -> usize { ... }
    fn aead_tag_length(alg: AeadAlgorithm) -> usize { ... }

}

The HpkeCrypto trait defines the necessary cryptographic functions used in the HPKE implementation.

Required Associated Types

type HpkePrng: RngCore + CryptoRng + HpkeTestRng

The PRNG implementation returned in HpkeCrypto::prng().

Required Methods

fn name() -> String

The name of the implementation.

fn supports_kdf(alg: KdfAlgorithm) -> Result<(), Error>

Returns an error if the KDF algorithm is not supported by this crypto provider.

fn supports_kem(alg: KemAlgorithm) -> Result<(), Error>

Returns an error if the KEM algorithm is not supported by this crypto provider.

fn supports_aead(alg: AeadAlgorithm) -> Result<(), Error>

Returns an error if the AEAD algorithm is not supported by this crypto provider.

fn prng() -> Self::HpkePrng

Get a stateful PRNG. Note that this will create a new PRNG state.

fn kdf_extract( alg: KdfAlgorithm, salt: &[u8], ikm: &[u8], ) -> Result<Vec<u8>, Error>

KDF Extract

fn kdf_expand( alg: KdfAlgorithm, prk: &[u8], info: &[u8], output_size: usize, ) -> Result<Vec<u8>, Error>

KDF Expand

fn dh(alg: KemAlgorithm, pk: &[u8], sk: &[u8]) -> Result<Vec<u8>, Error>

Diffie-Hellman

fn secret_to_public(alg: KemAlgorithm, sk: &[u8]) -> Result<Vec<u8>, Error>

Diffie-Hellman with the base (generate public key for secret key sk).

fn kem_key_gen( alg: KemAlgorithm, prng: &mut Self::HpkePrng, ) -> Result<(Vec<u8>, Vec<u8>), Error>

KEM key pair generation (encapsulation key, decapsulation key).

fn kem_key_gen_derand( alg: KemAlgorithm, seed: &[u8], ) -> Result<(Vec<u8>, Vec<u8>), Error>

KEM key pair generation (encapsulation key, decapsulation key) based on the seed.

fn kem_encaps( alg: KemAlgorithm, pk_r: &[u8], prng: &mut Self::HpkePrng, ) -> Result<(Vec<u8>, Vec<u8>), Error>

KEM encapsulation to pk_r (shared secret, ciphertext).

fn kem_decaps( alg: KemAlgorithm, ct: &[u8], sk_r: &[u8], ) -> Result<Vec<u8>, Error>

KEM decapsulation with sk_r. Returns the shared secret.

fn dh_validate_sk(alg: KemAlgorithm, sk: &[u8]) -> Result<Vec<u8>, Error>

Validate a secret key for its correctness.

fn aead_seal( alg: AeadAlgorithm, key: &[u8], nonce: &[u8], aad: &[u8], msg: &[u8], ) -> Result<Vec<u8>, Error>

AEAD encrypt.

fn aead_open( alg: AeadAlgorithm, key: &[u8], nonce: &[u8], aad: &[u8], msg: &[u8], ) -> Result<Vec<u8>, Error>

AEAD decrypt.

Provided Methods

fn kdf_digest_length(alg: KdfAlgorithm) -> usize

Get the length of the output digest.

fn aead_key_length(alg: AeadAlgorithm) -> usize

Get key length for AEAD.

Note that this function returns 0 for export only keys of unknown size.

fn aead_nonce_length(alg: AeadAlgorithm) -> usize

Get key length for AEAD.

Note that this function returns 0 for export only nonces of unknown size.

fn aead_tag_length(alg: AeadAlgorithm) -> usize

Get key length for AEAD.

Note that this function returns 0 for export only tags of unknown size.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.